requirements.txt

# AI Providers
# Core Framework
# Infrastructure
# Observability
# Testing
# Utilities
# Web Framework
asyncpg>=0.29.0
httpx>=0.27.0
kafka-python>=2.0.2
langchain-core==0.2.39
langchain-openai==0.1.23
langchain==0.2.16
langchain>=0.2.17
langgraph==0.2.27
langsmith==0.1.93
openai>=1.33.0
opentelemetry-api>=1.20.0
opentelemetry-exporter-prometheus>=0.41b0
opentelemetry-sdk>=1.20.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest>=7.4.0
python-dotenv>=1.0.0
redis>=5.0.0
streamlit>=1.30.0
